+++ backend/orchestrator/computeplan_pb2.py
@@ -16,7 +16,7 @@ _sym_db = _symbol_database.Default()
 from google.protobuf import timestamp_pb2 as google_dot_protobuf_dot_timestamp__pb2
 
 
-DESCRIPTOR = _descriptor_pool.Default().AddSerializedFile(b'\n\x11\x63omputeplan.proto\x12\x0corchestrator\x1a\x1fgoogle/protobuf/timestamp.proto\"\xce\x03\n\x0b\x43omputePlan\x12\x0b\n\x03key\x18\x01 \x01(\t\x12\r\n\x05owner\x18\x02 \x01(\t\x12\x15\n\rwaiting_count\x18\x08 \x01(\r\x12\x12\n\ntodo_count\x18\t \x01(\r\x12\x13\n\x0b\x64oing_count\x18\n \x01(\r\x12\x16\n\x0e\x63\x61nceled_count\x18\x0b \x01(\r\x12\x14\n\x0c\x66\x61iled_count\x18\x0c \x01(\r\x12\x12\n\ndone_count\x18\x03 \x01(\r\x12\x12\n\ntask_count\x18\x04 \x01(\r\x12/\n\x06status\x18\x05 \x01(\x0e\x32\x1f.orchestrator.ComputePlanStatus\x12\"\n\x1a\x64\x65lete_intermediary_models\x18\x06 \x01(\x08\x12\x31\n\rcreation_date\x18\x07 \x01(\x0b\x32\x1a.google.protobuf.Timestamp\x12\x0b\n\x03tag\x18\x10 \x01(\t\x12\x0c\n\x04name\x18\x13 \x01(\t\x12\x39\n\x08metadata\x18\x11 \x03(\x0b\x32\'.orchestrator.ComputePlan.MetadataEntry\x1a/\n\rMetadataEntry\x12\x0b\n\x03key\x18\x01 \x01(\t\x12\r\n\x05value\x18\x02 \x01(\t:\x02\x38\x01\"\xcb\x01\n\x0eNewComputePlan\x12\x0b\n\x03key\x18\x01 \x01(\t\x12\x0b\n\x03tag\x18\x10 \x01(\t\x12\x0c\n\x04name\x18\x13 \x01(\t\x12<\n\x08metadata\x18\x11 \x03(\x0b\x32*.orchestrator.NewComputePlan.MetadataEntry\x12\"\n\x1a\x64\x65lete_intermediary_models\x18\x12 \x01(\x08\x1a/\n\rMetadataEntry\x12\x0b\n\x03key\x18\x01 \x01(\t\x12\r\n\x05value\x18\x02 \x01(\t:\x02\x38\x01\"\"\n\x13GetComputePlanParam\x12\x0b\n\x03key\x18\x01 \x01(\t\"T\n\x14\x41pplyPlanActionParam\x12\x0b\n\x03key\x18\x01 \x01(\t\x12/\n\x06\x61\x63tion\x18\x02 \x01(\x0e\x32\x1f.orchestrator.ComputePlanAction\"\x19\n\x17\x41pplyPlanActionResponse\" \n\x0fPlanQueryFilter\x12\r\n\x05owner\x18\x01 \x01(\t\"g\n\x0fQueryPlansParam\x12\x12\n\npage_token\x18\x01 \x01(\t\x12\x11\n\tpage_size\x18\x02 \x01(\r\x12-\n\x06\x66ilter\x18\x03 \x01(\x0b\x32\x1d.orchestrator.PlanQueryFilter\"W\n\x12QueryPlansResponse\x12(\n\x05plans\x18\x01 \x03(\x0b\x32\x19.orchestrator.ComputePlan\x12\x17\n\x0fnext_page_token\x18\x02 \x01(\t*\xba\x01\n\x11\x43omputePlanStatus\x12\x17\n\x13PLAN_STATUS_UNKNOWN\x10\x00\x12\x17\n\x13PLAN_STATUS_WAITING\x10\x01\x12\x14\n\x10PLAN_STATUS_TODO\x10\x02\x12\x15\n\x11PLAN_STATUS_DOING\x10\x03\x12\x14\n\x10PLAN_STATUS_DONE\x10\x04\x12\x18\n\x14PLAN_STATUS_CANCELED\x10\x05\x12\x16\n\x12PLAN_STATUS_FAILED\x10\x06*F\n\x11\x43omputePlanAction\x12\x17\n\x13PLAN_ACTION_UNKNOWN\x10\x00\x12\x18\n\x14PLAN_ACTION_CANCELED\x10\x01\x32\xd3\x02\n\x12\x43omputePlanService\x12G\n\x0cRegisterPlan\x12\x1c.orchestrator.NewComputePlan\x1a\x19.orchestrator.ComputePlan\x12G\n\x07GetPlan\x12!.orchestrator.GetComputePlanParam\x1a\x19.orchestrator.ComputePlan\x12\\\n\x0f\x41pplyPlanAction\x12\".orchestrator.ApplyPlanActionParam\x1a%.orchestrator.ApplyPlanActionResponse\x12M\n\nQueryPlans\x12\x1d.orchestrator.QueryPlansParam\x1a .orchestrator.QueryPlansResponseB)Z\'github.com/owkin/orchestrator/lib/assetb\x06proto3')
+DESCRIPTOR = _descriptor_pool.Default().AddSerializedFile(b'\n\x11\x63omputeplan.proto\x12\x0corchestrator\x1a\x1fgoogle/protobuf/timestamp.proto\"\xce\x03\n\x0b\x43omputePlan\x12\x0b\n\x03key\x18\x01 \x01(\t\x12\r\n\x05owner\x18\x02 \x01(\t\x12\x15\n\rwaiting_count\x18\x08 \x01(\r\x12\x12\n\ntodo_count\x18\t \x01(\r\x12\x13\n\x0b\x64oing_count\x18\n \x01(\r\x12\x16\n\x0e\x63\x61nceled_count\x18\x0b \x01(\r\x12\x14\n\x0c\x66\x61iled_count\x18\x0c \x01(\r\x12\x12\n\ndone_count\x18\x03 \x01(\r\x12\x12\n\ntask_count\x18\x04 \x01(\r\x12/\n\x06status\x18\x05 \x01(\x0e\x32\x1f.orchestrator.ComputePlanStatus\x12\"\n\x1a\x64\x65lete_intermediary_models\x18\x06 \x01(\x08\x12\x31\n\rcreation_date\x18\x07 \x01(\x0b\x32\x1a.google.protobuf.Timestamp\x12\x0b\n\x03tag\x18\x10 \x01(\t\x12\x0c\n\x04name\x18\x13 \x01(\t\x12\x39\n\x08metadata\x18\x11 \x03(\x0b\x32\'.orchestrator.ComputePlan.MetadataEntry\x1a/\n\rMetadataEntry\x12\x0b\n\x03key\x18\x01 \x01(\t\x12\r\n\x05value\x18\x02 \x01(\t:\x02\x38\x01\"\xcb\x01\n\x0eNewComputePlan\x12\x0b\n\x03key\x18\x01 \x01(\t\x12\x0b\n\x03tag\x18\x10 \x01(\t\x12\x0c\n\x04name\x18\x13 \x01(\t\x12<\n\x08metadata\x18\x11 \x03(\x0b\x32*.orchestrator.NewComputePlan.MetadataEntry\x12\"\n\x1a\x64\x65lete_intermediary_models\x18\x12 \x01(\x08\x1a/\n\rMetadataEntry\x12\x0b\n\x03key\x18\x01 \x01(\t\x12\r\n\x05value\x18\x02 \x01(\t:\x02\x38\x01\"\"\n\x13GetComputePlanParam\x12\x0b\n\x03key\x18\x01 \x01(\t\"T\n\x14\x41pplyPlanActionParam\x12\x0b\n\x03key\x18\x01 \x01(\t\x12/\n\x06\x61\x63tion\x18\x02 \x01(\x0e\x32\x1f.orchestrator.ComputePlanAction\"\x19\n\x17\x41pplyPlanActionResponse\" \n\x0fPlanQueryFilter\x12\r\n\x05owner\x18\x01 \x01(\t\"g\n\x0fQueryPlansParam\x12\x12\n\npage_token\x18\x01 \x01(\t\x12\x11\n\tpage_size\x18\x02 \x01(\r\x12-\n\x06\x66ilter\x18\x03 \x01(\x0b\x32\x1d.orchestrator.PlanQueryFilter\"W\n\x12QueryPlansResponse\x12(\n\x05plans\x18\x01 \x03(\x0b\x32\x19.orchestrator.ComputePlan\x12\x17\n\x0fnext_page_token\x18\x02 \x01(\t*\xd1\x01\n\x11\x43omputePlanStatus\x12\x17\n\x13PLAN_STATUS_UNKNOWN\x10\x00\x12\x17\n\x13PLAN_STATUS_WAITING\x10\x01\x12\x14\n\x10PLAN_STATUS_TODO\x10\x02\x12\x15\n\x11PLAN_STATUS_DOING\x10\x03\x12\x14\n\x10PLAN_STATUS_DONE\x10\x04\x12\x18\n\x14PLAN_STATUS_CANCELED\x10\x05\x12\x16\n\x12PLAN_STATUS_FAILED\x10\x06\x12\x15\n\x11PLAN_STATUS_EMPTY\x10\x07*F\n\x11\x43omputePlanAction\x12\x17\n\x13PLAN_ACTION_UNKNOWN\x10\x00\x12\x18\n\x14PLAN_ACTION_CANCELED\x10\x01\x32\xd3\x02\n\x12\x43omputePlanService\x12G\n\x0cRegisterPlan\x12\x1c.orchestrator.NewComputePlan\x1a\x19.orchestrator.ComputePlan\x12G\n\x07GetPlan\x12!.orchestrator.GetComputePlanParam\x1a\x19.orchestrator.ComputePlan\x12\\\n\x0f\x41pplyPlanAction\x12\".orchestrator.ApplyPlanActionParam\x1a%.orchestrator.ApplyPlanActionResponse\x12M\n\nQueryPlans\x12\x1d.orchestrator.QueryPlansParam\x1a .orchestrator.QueryPlansResponseB)Z\'github.com/owkin/orchestrator/lib/assetb\x06proto3')
 
 _COMPUTEPLANSTATUS = DESCRIPTOR.enum_types_by_name['ComputePlanStatus']
 ComputePlanStatus = enum_type_wrapper.EnumTypeWrapper(_COMPUTEPLANSTATUS)
@@ -29,6 +29,7 @@ PLAN_STATUS_DOING = 3
 PLAN_STATUS_DONE = 4
 PLAN_STATUS_CANCELED = 5
 PLAN_STATUS_FAILED = 6
+PLAN_STATUS_EMPTY = 7
 PLAN_ACTION_UNKNOWN = 0
 PLAN_ACTION_CANCELED = 1
 
@@ -125,9 +126,9 @@ if _descriptor._USE_C_DESCRIPTORS == False:
   _NEWCOMPUTEPLAN_METADATAENTRY._options = None
   _NEWCOMPUTEPLAN_METADATAENTRY._serialized_options = b'8\001'
   _COMPUTEPLANSTATUS._serialized_start=1117
-  _COMPUTEPLANSTATUS._serialized_end=1303
-  _COMPUTEPLANACTION._serialized_start=1305
-  _COMPUTEPLANACTION._serialized_end=1375
+  _COMPUTEPLANSTATUS._serialized_end=1326
+  _COMPUTEPLANACTION._serialized_start=1328
+  _COMPUTEPLANACTION._serialized_end=1398
   _COMPUTEPLAN._serialized_start=69
   _COMPUTEPLAN._serialized_end=531
   _COMPUTEPLAN_METADATAENTRY._serialized_start=484
@@ -148,6 +149,6 @@ if _descriptor._USE_C_DESCRIPTORS == False:
   _QUERYPLANSPARAM._serialized_end=1025
   _QUERYPLANSRESPONSE._serialized_start=1027
   _QUERYPLANSRESPONSE._serialized_end=1114
-  _COMPUTEPLANSERVICE._serialized_start=1378
-  _COMPUTEPLANSERVICE._serialized_end=1717
+  _COMPUTEPLANSERVICE._serialized_start=1401
+  _COMPUTEPLANSERVICE._serialized_end=1740
 # @@protoc_insertion_point(module_scope)